Implementation/IT/requirements.txt
asgiref==3.4.1
Django==4.0
sqlparse==0.4.2
psycopg2==2.9.2
django-rest-framework==0.1.0
djangorestframework~=3.13.1
django-cors-headers==3.10.1
djoser==2.1.0